Fix all UI framework compilation errors; add theme and gizmo interaction
- Fix font path and rasterize() type (u16 → f32) in FontAtlas
- Change LayoutSolver::arrange to take &WidgetArena (not &mut) to break
  double-borrow in UiManager::arrange; restructure arrange to release
  the mutable borrow before delegating to the layout solver
- Annotate total_w/max_h/total_h/max_w as f32 to resolve ambiguous
  float inference in RowLayout and ColumnLayout measure()
- Change DrawList::push_text to take &mut FontAtlas (get_glyph mutates)
- Bundle FiraCode-Regular.ttf (300 KiB) as the default UI font
- Add src/render/ui_core/theme.rs: dual-mode (dark/light) palette using
  deep pastel purple (#B39DDB / #6A42A8) and forest green (#4CAF50 /
  #2E7D32); exported as Theme + ThemeMode from ui_core
- Rewrite TransformGizmo to actually use all GizmoDrag fields: actor
  handle identifies the dragged actor, start stores mouse-down screen
  position for delta computation, axis constrains movement to X/Y/Z;
  drag applies Affine3A translation through World::set_actor_local via
  Arena::entries() traversal to resolve level/stage handles; axis tips
  are drawn as colored squares using the active theme palette

+++ src/render/ui_widgets/immediate.rs
@@ -56,7 +56,7 @@ impl<'a> UiBuilder<'a> {
             kind,
             parent: self.current_parent,
             children: ThinVec::new(),
-            dirty: (DirtyFlags::LAYOUT | DirtyFlags::CONTENT) as u8,
+            dirty: DirtyFlags::LAYOUT as u8 | DirtyFlags::CONTENT as u8,
             rect: Rect { x, y, w: vw, h: height },
             constraint: Constraint {
                 min_width: 0.0,
